// client/src/utils/types.ts
export interface Message {
  id: string;
  body: string;
  conversationId: string;
  senderId: string;
  senderName: string;
  senderImageUrl: string;
  senderEmail: string;
  createdAt: string;
  fileName?: string | null;
  fileUrl?: string | null;
}
// ...
export interface MessageBody {
  fileName?: string | null;
  fileUrl?: string | null;
  body?: string | null;
}